[storage] trivial: extract transaction store iterators to separate file
