+++ storage/aptosdb/src/transaction_store/mod.rs
@@ -20,9 +20,14 @@ use aptos_types::{
     transaction::{Transaction, Version},
     write_set::WriteSet,
 };
-use schemadb::iterator::SchemaIterator;
+use iterators::{AccountTransactionVersionIter, TransactionIter};
 use schemadb::{ReadOptions, SchemaBatch, DB};
 use std::sync::Arc;
+
+mod iterators;
+
+#[cfg(test)]
+mod test;
 
 #[derive(Clone, Debug)]
 pub struct TransactionStore {
@@ -82,16 +87,14 @@ impl TransactionStore {
             .db
             .iter::<TransactionByAccountSchema>(ReadOptions::default())?;
         iter.seek(&(address, min_seq_num))?;
-        Ok(AccountTransactionVersionIter {
-            inner: iter,
+        Ok(AccountTransactionVersionIter::new(
+            iter,
             address,
-            expected_next_seq_num: None,
-            end_seq_num: min_seq_num
+            min_seq_num
                 .checked_add(num_versions)
                 .ok_or_else(|| format_err!("too many transactions requested"))?,
-            prev_version: None,
             ledger_version,
-        })
+        ))
     }
 
     /// Get signed transaction given `version`
@@ -109,13 +112,7 @@ impl TransactionStore {
     ) -> Result<TransactionIter> {
         let mut iter = self.db.iter::<TransactionSchema>(ReadOptions::default())?;
         iter.seek(&start_version)?;
-        Ok(TransactionIter {
-            inner: iter,
-            expected_next_version: start_version,
-            end_version: start_version
-                .checked_add(num_transactions as u64)
-                .ok_or_else(|| format_err!("too many transactions requested"))?,
-        })
+        TransactionIter::new(iter, start_version, num_transactions)
     }
 
     /// Save signed transaction at `version`
@@ -198,7 +195,7 @@ impl TransactionStore {
         &self,
         transactions: &[Transaction],
         db_batch: &mut SchemaBatch,
-    ) -> anyhow::Result<()> {
+    ) -> Result<()> {
         for transaction in transactions {
             db_batch.delete::<TransactionByHashSchema>(&transaction.hash())?;
         }
@@ -210,7 +207,7 @@ impl TransactionStore {
         &self,
         transactions: &[Transaction],
         db_batch: &mut SchemaBatch,
-    ) -> anyhow::Result<()> {
+    ) -> Result<()> {
         for transaction in transactions {
             if let Transaction::UserTransaction(txn) = transaction {
                 db_batch
@@ -226,7 +223,7 @@ impl TransactionStore {
         begin: Version,
         end: Version,
         db_batch: &mut SchemaBatch,
-    ) -> anyhow::Result<()> {
+    ) -> Result<()> {
         for version in begin..end {
             db_batch.delete::<TransactionSchema>(&version)?;
         }
@@ -239,7 +236,7 @@ impl TransactionStore {
         begin: Version,
         end: Version,
         db_batch: &mut SchemaBatch,
-    ) -> anyhow::Result<()> {
+    ) -> Result<()> {
         for version in begin..end {
             db_batch.delete::<TransactionInfoSchema>(&version)?;
         }
@@ -261,7 +258,7 @@ impl TransactionStore {
         begin: Version,
         end: Version,
         db_batch: &mut SchemaBatch,
-    ) -> anyhow::Result<()> {
+    ) -> Result<()> {
         for version_to_delete in begin..end {
             // The even version will be pruned in the iteration of version + 1.
             if version_to_delete % 2 == 0 {
@@ -311,114 +308,10 @@ impl TransactionStore {
         begin: Version,
         end: Version,
         db_batch: &mut SchemaBatch,
-    ) -> anyhow::Result<()> {
+    ) -> Result<()> {
         for version in begin..end {
             db_batch.delete::<WriteSetSchema>(&version)?;
         }
         Ok(())
     }
 }
-
-pub struct TransactionIter<'a> {
-    inner: SchemaIterator<'a, TransactionSchema>,
-    expected_next_version: Version,
-    end_version: Version,
-}
-
-impl<'a> TransactionIter<'a> {
-    fn next_impl(&mut self) -> Result<Option<Transaction>> {
-        if self.expected_next_version >= self.end_version {
-            return Ok(None);
-        }
-
-        let ret = match self.inner.next().transpose()? {
-            Some((version, transaction)) => {
-                ensure!(
-                    version == self.expected_next_version,
-                    "Transaction versions are not consecutive.",
-                );
-                self.expected_next_version += 1;
-                Some(transaction)
-            }
-            None => None,
-        };
-
-        Ok(ret)
-    }
-}
-
-impl<'a> Iterator for TransactionIter<'a> {
-    type Item = Result<Transaction>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.next_impl().transpose()
-    }
-}
-
-pub struct AccountTransactionVersionIter<'a> {
-    inner: SchemaIterator<'a, TransactionByAccountSchema>,
-    address: AccountAddress,
-    expected_next_seq_num: Option<u64>,
-    end_seq_num: u64,
-    prev_version: Option<Version>,
-    ledger_version: Version,
-}
-
-impl<'a> AccountTransactionVersionIter<'a> {
-    fn next_impl(&mut self) -> Result<Option<(u64, Version)>> {
-        Ok(match self.inner.next().transpose()? {
-            Some(((address, seq_num), version)) => {
-                // No more transactions sent by this account.
-                if address != self.address {
-                    return Ok(None);
-                }
-                if seq_num >= self.end_seq_num {
-                    return Ok(None);
-                }
-
-                // Ensure seq_num_{i+1} == seq_num_{i} + 1
-                if let Some(expected_seq_num) = self.expected_next_seq_num {
-                    ensure!(
-                        seq_num == expected_seq_num,
-                        "DB corruption: account transactions sequence numbers are not contiguous: \
-                     actual: {}, expected: {}",
-                        seq_num,
-                        expected_seq_num,
-                    );
-                };
-
-                // Ensure version_{i+1} > version_{i}
-                if let Some(prev_version) = self.prev_version {
-                    ensure!(
-                        prev_version < version,
-                        "DB corruption: account transaction versions are not strictly increasing: \
-                         previous version: {}, current version: {}",
-                        prev_version,
-                        version,
-                    );
-                }
-
-                // No more transactions (in this view of the ledger).
-                if version > self.ledger_version {
-                    return Ok(None);
-                }
-
-                self.expected_next_seq_num = Some(seq_num + 1);
-                self.prev_version = Some(version);
-                Some((seq_num, version))
-            }
-            None => None,
-        })
-    }
-}
-
-impl<'a> Iterator for AccountTransactionVersionIter<'a> {
-    type Item = Result<(u64, Version)>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.next_impl().transpose()
-    }
-}
-
-#[cfg(test)]
-mod test;
